Feat: modularize upload and review pages
