+++ client/src/pages/Review/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./Review";
+export * from "./Review";
