+++ client/src/pages/Upload/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./Upload";
+export * from "./Upload";
